refactor(#129): split trajectory optimizer modules
