+++ src/drake_models/optimization/drake_trajectory_solver.py
@@ -1,0 +1,243 @@
+"""Drake-backed direct transcription trajectory solver."""
+
+from __future__ import annotations
+
+import numpy as np
+
+from drake_models.optimization.exercise_objectives import ExerciseObjective
+from drake_models.optimization.trajectory_types import (
+    TrajectoryConfig,
+    TrajectoryResult,
+)
+from drake_models.shared.contracts.preconditions import require_finite
+
+
+def _build_drake_plant(sdf_string: str, dt: float) -> object:
+    """Load *sdf_string* into a finalised Drake MultibodyPlant."""
+    from pydrake.multibody.parsing import Parser
+    from pydrake.multibody.plant import AddMultibodyPlantSceneGraph
+    from pydrake.systems.framework import DiagramBuilder
+
+    builder = DiagramBuilder()
+    plant, _scene_graph = AddMultibodyPlantSceneGraph(builder, time_step=dt)
+    parser = Parser(plant)
+    parser.AddModelsFromString(sdf_string, "sdf")
+    plant.Finalize()
+    return plant
+
+
+def _add_control_costs(prog: object, u: object, n_steps: int, weight: float) -> None:
+    """Add per-timestep quadratic control costs to *prog*."""
+    n_u = u.shape[1]  # type: ignore[attr-defined]
+    for k in range(n_steps):
+        prog.AddQuadraticCost(  # type: ignore[attr-defined]
+            weight * np.eye(n_u),
+            np.zeros(n_u),
+            u[k],  # type: ignore[index]
+        )
+
+
+def _add_integration_constraints(
+    prog: object,
+    q: object,
+    v: object,
+    dt: float,
+    n_steps: int,
+) -> int:
+    """Add semi-implicit Euler integration constraints."""
+    n_q = q.shape[1]  # type: ignore[attr-defined]
+    n_v = v.shape[1]  # type: ignore[attr-defined]
+    offset = n_q - n_v
+    added = 0
+    for k in range(n_steps - 1):
+        for j in range(n_v):
+            prog.AddLinearEqualityConstraint(  # type: ignore[attr-defined]
+                q[k + 1, offset + j]  # type: ignore[index]
+                - q[k, offset + j]  # type: ignore[index]
+                - dt * v[k + 1, j]  # type: ignore[index]
+                == 0
+            )
+            added += 1
+    return added
+
+
+def _add_dynamics_constraints(
+    prog: object,
+    plant: object,
+    q: object,
+    v: object,
+    u: object,
+    dt: float,
+    n_steps: int,
+) -> int:
+    """Add per-knot manipulator-equation dynamics constraints to *prog*."""
+    n_q = q.shape[1]  # type: ignore[attr-defined]
+    n_v = v.shape[1]  # type: ignore[attr-defined]
+    n_u = u.shape[1]  # type: ignore[attr-defined]
+    context = plant.CreateDefaultContext()  # type: ignore[attr-defined]
+    actuation = plant.MakeActuationMatrix()  # type: ignore[attr-defined]
+
+    def _residual(vars_flat: np.ndarray) -> np.ndarray:
+        qk = vars_flat[:n_q]
+        vk = vars_flat[n_q : n_q + n_v]
+        vkp1 = vars_flat[n_q + n_v : n_q + 2 * n_v]
+        uk = vars_flat[n_q + 2 * n_v : n_q + 2 * n_v + n_u]
+        plant.SetPositions(context, qk)  # type: ignore[attr-defined]
+        plant.SetVelocities(context, vk)  # type: ignore[attr-defined]
+        mass = plant.CalcMassMatrix(context)  # type: ignore[attr-defined]
+        bias = plant.CalcBiasTerm(context)  # type: ignore[attr-defined]
+        gravity = plant.CalcGravityGeneralizedForces(context)  # type: ignore[attr-defined]
+        vdot = (vkp1 - vk) / dt
+        return mass @ vdot + bias - gravity - actuation @ uk
+
+    lb = np.zeros(n_v)
+    ub = np.zeros(n_v)
+    for k in range(n_steps - 1):
+        vars_k = np.concatenate(
+            [
+                q[k],  # type: ignore[index]
+                v[k],  # type: ignore[index]
+                v[k + 1],  # type: ignore[index]
+                u[k],  # type: ignore[index]
+            ]
+        )
+        prog.AddConstraint(_residual, lb=lb, ub=ub, vars=vars_k)  # type: ignore[attr-defined]
+    return n_steps - 1
+
+
+def _add_initial_state_constraint(
+    prog: object,
+    q: object,
+    v: object,
+    q0: np.ndarray,
+    v0: np.ndarray,
+) -> int:
+    """Pin the first knot point to the supplied initial state."""
+    n_q = q.shape[1]  # type: ignore[attr-defined]
+    n_v = v.shape[1]  # type: ignore[attr-defined]
+    for j in range(n_q):
+        prog.AddLinearEqualityConstraint(  # type: ignore[attr-defined]
+            q[0, j] == float(q0[j])  # type: ignore[index]
+        )
+    for j in range(n_v):
+        prog.AddLinearEqualityConstraint(  # type: ignore[attr-defined]
+            v[0, j] == float(v0[j])  # type: ignore[index]
+        )
+    return n_q + n_v
+
+
+def _add_joint_and_actuator_bounds(
+    prog: object,
+    plant: object,
+    q: object,
+    u: object,
+    n_steps: int,
+) -> int:
+    """Apply per-knot position limits and actuator effort limits."""
+    q_lower = plant.GetPositionLowerLimits()  # type: ignore[attr-defined]
+    q_upper = plant.GetPositionUpperLimits()  # type: ignore[attr-defined]
+    u_lower = plant.GetEffortLowerLimits()  # type: ignore[attr-defined]
+    u_upper = plant.GetEffortUpperLimits()  # type: ignore[attr-defined]
+
+    q_lower = np.where(np.isfinite(q_lower), q_lower, -1e9)
+    q_upper = np.where(np.isfinite(q_upper), q_upper, 1e9)
+    u_lower = np.where(np.isfinite(u_lower), u_lower, -1e9)
+    u_upper = np.where(np.isfinite(u_upper), u_upper, 1e9)
+
+    added = 0
+    for k in range(n_steps):
+        prog.AddBoundingBoxConstraint(q_lower, q_upper, q[k])  # type: ignore[attr-defined,index]
+        prog.AddBoundingBoxConstraint(u_lower, u_upper, u[k])  # type: ignore[attr-defined,index]
+        added += 2
+    return added
+
+
+def _add_phase_tracking_costs(
+    prog: object,
+    q: object,
+    objective: ExerciseObjective,
+    n_q: int,
+    n_steps: int,
+    state_weight: float,
+    terminal_weight: float,
+) -> None:
+    """Add phase-tracking quadratic costs to *prog*."""
+    joint_names = objective.joint_names()
+    for phase in objective.phases:
+        k = int(phase.time_fraction * (n_steps - 1))
+        target = np.zeros(n_q)
+        for jname, angle in phase.joint_angles.items():
+            if jname in joint_names:
+                idx = joint_names.index(jname)
+                if idx < n_q:
+                    target[idx] = angle
+        weight = terminal_weight if phase is objective.phases[-1] else state_weight
+        prog.AddQuadraticCost(  # type: ignore[attr-defined]
+            weight * np.eye(n_q),
+            -weight * target,
+            q[k],  # type: ignore[index]
+        )
+
+
+def _build_drake_program(
+    plant: object,
+    objective: ExerciseObjective,
+    config: TrajectoryConfig,
+) -> tuple[object, object, object, object]:
+    """Construct the MathematicalProgram with variables, costs, and constraints."""
+    from pydrake.solvers import MathematicalProgram
+
+    n_q = plant.num_positions()  # type: ignore[attr-defined]
+    n_v = plant.num_velocities()  # type: ignore[attr-defined]
+    n_u = plant.num_actuators()  # type: ignore[attr-defined]
+    n_steps = config.n_timesteps
+
+    prog = MathematicalProgram()
+    q = prog.NewContinuousVariables(n_steps, n_q, "q")
+    v = prog.NewContinuousVariables(n_steps, n_v, "v")
+    u = prog.NewContinuousVariables(n_steps, n_u, "u")
+
+    _add_control_costs(prog, u, n_steps, config.control_weight)
+    _add_phase_tracking_costs(
+        prog,
+        q,
+        objective,
+        n_q,
+        n_steps,
+        config.state_weight,
+        config.terminal_weight,
+    )
+    _add_integration_constraints(prog, q, v, config.dt, n_steps)
+    _add_dynamics_constraints(prog, plant, q, v, u, config.dt, n_steps)
+    _add_joint_and_actuator_bounds(prog, plant, q, u, n_steps)
+
+    context = plant.CreateDefaultContext()  # type: ignore[attr-defined]
+    q0 = np.asarray(plant.GetPositions(context))  # type: ignore[attr-defined]
+    v0 = np.zeros(n_v)
+    _add_initial_state_constraint(prog, q, v, q0, v0)
+    return prog, q, v, u
+
+
+def solve_with_drake(
+    sdf_string: str,
+    objective: ExerciseObjective,
+    config: TrajectoryConfig,
+) -> TrajectoryResult:
+    """Solve trajectory optimization using Drake's MathematicalProgram."""
+    from pydrake.solvers import Solve
+
+    require_finite(np.array([config.dt, config.total_time]), "config timing")
+    plant = _build_drake_plant(sdf_string, config.dt)
+    prog, q, v, u = _build_drake_program(plant, objective, config)
+    result = Solve(prog)
+    time = np.linspace(0.0, config.total_time, config.n_timesteps)
+
+    return TrajectoryResult(
+        joint_positions=result.GetSolution(q),
+        joint_velocities=result.GetSolution(v),
+        joint_torques=result.GetSolution(u),
+        time=time,
+        cost=result.get_optimal_cost(),
+        converged=result.is_success(),
+        iterations=config.max_iterations,
+    )
